fix: copy texts whenever mutating them, to avoid concurrent modification exceptions when messages are being broadcast to many players with Server Translations API installed
PlayerList#broadcastSystemMessage will send the same text in a packet to all players. Impersonate will modify the text's siblings before the packet is sent, on the main thread. Server Translations API then modifies the text's siblings while the packet is sent, on the network thread. With enough players, there is a chance that both happen at the same time, causing a concurrent modification exception.